Update react version, move to using a bundler.
