+++ webpack.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  entry: "./src/main.js",
+  output: {
+    filename: "build.js",
+    path: `${__dirname}/data`
+  }
+};
